requirements.txt
hvac
SQLAlchemy
psutil
numpy
opencv-python-headless
PyYAML
fastapi
uvicorn
prometheus-client
python-dotenv
httpx
gnupg
loguru
